requirements.txt
click>=7,<8
pyzmq>=18,<19
PyYAML>=3
colorful>=0.5,<1.0
